refactor: Use nagation glob to keep all tests under the tests dir
