+++ .travis.yml
@@ -22,10 +22,9 @@ install:
   - yarn install --no-lockfile --non-interactive
 
 before_script:
-  - lerna bootstrap
+  - yarn run bootstrap
   - yarn run build
 
 script:
   - commitlint-travis
-  - yarn run lint
-  - yarn run test
+  - yarn run test:ci
