+++ packages/@pollyjs/core/tests/browser/integration/adapters-test.js
@@ -1,6 +1,6 @@
 import 'formdata-polyfill';
 
-import { setupMocha as setupPolly } from '../../src';
+import { setupMocha as setupPolly } from '../../../src';
 import * as setupFetch from '../helpers/setup-fetch';
 import File from '../helpers/file';
 import Configs from './configs';
