+++ packages/@pollyjs/ember/package.json
@@ -34,7 +34,7 @@
     "lint": "eslint ./*.js addon addon-test-support app config lib server test-support tests",
     "start": "ember serve",
     "format": "prettier --config ../../../.prettier.js --write **/*.js",
-    "test": "ember test ci",
+    "test": "ember test ci --port=7300",
     "clean": "rimraf dist tmp"
   },
   "dependencies": {
