+++ packages/@pollyjs/persister-fs/tests/node/unit/fs-persister-test.js
@@ -1,4 +1,4 @@
-import FSPersister from '../../src';
+import FSPersister from '../../../src';
 
 describe('Unit | FS Persister', function() {
   it('should exist', function() {
